[move] : move to mypage's components
